[NO-ISSUE] Refactor for Facade, Services and DTO.

+++ src/main/java/pl/kf/itjobsearcher/business/offer/dto/CreateOfferCommand.java
@@ -4,6 +4,7 @@ import lombok.Builder;
 import pl.kf.itjobsearcher.business.offer.domain.model.ContractType;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 
 @Builder
@@ -11,6 +12,6 @@ public record CreateOfferCommand(
         String title,
         String description,
         List<ContractType> contractTypes,
-        BigDecimal minSalary,
-        BigDecimal maxSalary
+        BigInteger minSalary,
+        BigInteger maxSalary
 ) { }
